Cherry-pick 4cac7925aca4. https://bugs.webkit.org/show_bug.cgi?id=272787
    [JSC] Mitigate null UnlinkedMetadataTable pointer in CodeBlock destructor
    https://bugs.webkit.org/show_bug.cgi?id=272787
    rdar://121747906

    Reviewed by Yusuke Suzuki.

    Attempts to fix a rare bug where the UnlinkedMetadataTable pointer accessed
    in the CodeBlock destructor can become null. We think this may be due to a
    series of thread-unsafe reference count operations that might allow the
    destructor to happen twice, perhaps simultaneously on two threads. This
    patch attempts to mitigate this by:

     1. Making UnlinkedMetadataTable and MetadataTable thread-safe refcounted.

     2. Checking for the presence of a null UnlinkedMetadataTable pointer in the
        appropriate functions, and attempting to handle it nonfatally. This means
        we skip updating the didOptimize state in the CodeBlock destructor, and
        that we intentionally leak MetadataTables if they have this null pointer.

    * Source/JavaScriptCore/bytecode/CodeBlock.cpp:
    (JSC::CodeBlock::~CodeBlock):
    * Source/JavaScriptCore/bytecode/MetadataTable.cpp:
    (JSC::MetadataTable::destroy):
    (JSC::MetadataTable::sizeInBytesForGC):
    * Source/JavaScriptCore/bytecode/MetadataTable.h:
    (JSC::MetadataTable::forEachValueProfile):
    (JSC::MetadataTable::valueProfileForOffset):
    (JSC::MetadataTable::deref):
    (JSC::MetadataTable::unlinkedMetadata const):
    (JSC::MetadataTable::totalSize const):
    * Source/JavaScriptCore/bytecode/UnlinkedMetadataTable.h:

    Canonical link: https://commits.webkit.org/272448.937@safari-7618-branch

Canonical link: https://commits.webkit.org/274313.249@webkitglib/2.44

+++ Source/JavaScriptCore/bytecode/MetadataTable.cpp
@@ -61,8 +61,16 @@ MetadataTable::~MetadataTable()
 
 void MetadataTable::destroy(MetadataTable* table)
 {
-    Ref<UnlinkedMetadataTable> unlinkedMetadata = WTFMove(table->linkingData().unlinkedMetadata);
+    RefPtr<UnlinkedMetadataTable> unlinkedMetadata = WTFMove(table->linkingData().unlinkedMetadata);
+
     table->~MetadataTable();
+
+    // FIXME: This check should really not be necessary, see https://webkit.org/b/272787
+    if (UNLIKELY(!unlinkedMetadata)) {
+        ASSERT_NOT_REACHED();
+        return;
+    }
+
     // Since UnlinkedMetadata::unlink frees the underlying memory of MetadataTable.
     // We need to destroy LinkingData before calling it.
     unlinkedMetadata->unlink(*table);
@@ -70,7 +78,7 @@ void MetadataTable::destroy(MetadataTable* table)
 
 size_t MetadataTable::sizeInBytesForGC()
 {
-    return unlinkedMetadata().sizeInBytesForGC(*this);
+    return unlinkedMetadata()->sizeInBytesForGC(*this);
 }
 
 void MetadataTable::validate() const
